% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USIG_geocode.R
\name{USIG_geocode}
\alias{USIG_geocode}
\title{USIG_geocode
Geocoder basado en el servicio de USIG GCABA (http://usig.buenosaires.gob.ar/)
uso: USIG_geocode(direccion)}
\usage{
USIG_geocode(address)
}
\arguments{
\item{direccion}{Una dirección dentro de los límites de la Región Urbana Metropolitana de Buenos Aires}
}
\description{
USIG_geocode
Geocoder basado en el servicio de USIG GCABA (http://usig.buenosaires.gob.ar/)
uso: USIG_geocode(direccion)
}
\examples{

Ejemplos:
USIG_geocode("9 de julio y belgrano")

resultado

              address              address_normalised        lat        lng
9 de julio y belgrano 9 DE JULIO AV. y BELGRANO, CABA -34.613090 -58.381226

Si se trata de una dirección fuera de la Ciudad Autónoma de Buenos Aires, explicitar el municipio o partido
USIG_geocode("9 de julio y belgrano, temperley")

 resultado

               address              address_normalised        lat        lng
9 de julio y belgrano 9 DE JULIO AV. y BELGRANO, CABA -34.613090 -58.381226
}
